organize files, run linters, edit readme file
