+++ README.md
@@ -1,2 +1,47 @@
-# js-testing-prac
-Some practice on introduction to testing
+# JavaScript Testing Practice
+
+[![Contributors][contributors-shield]][contributors-url]
+[![Forks][forks-shield]][forks-url]
+[![Stargazers][stars-shield]][stars-url]
+[![Issues][issues-shield]][issues-url]
+
+## About the project
+
+This project contains various scripts and then tests them using the Jest library.
+
+## Features
+
+Some of the features of this project include:
+
+- JavaScript scripts
+- Test files
+
+## Technologies Used
+
+This project utilizes:
+
+- JavaScript
+- Jest
+
+## Contributors
+
+- Alvaro Paz
+  - Linkedin: [@alvaropaz](https://linkedin.com/in/alvaropaz/)
+  - GitHub: [@alvaropazz](https://github.com/alvaropazz)
+  - Twitter: [@alvaroipazr](https://twitter.com/alvaroipazr)
+
+- Jay Maina
+  - LinkedIn : [Jay Maina](https://www.linkedin.com/in/jay-maina)
+  - GitHub: [@JayKowski](https://github.com/JayKowski)
+  - Twitter: [@Kowskijayy](https://twitter.com/Kowskijayy)
+  
+<!-- MARKDOWN LINKS & IMAGES -->
+
+[contributors-shield]: https://img.shields.io/github/contributors/JayKowski/js-testing-prac.svg?style=flat-square
+[contributors-url]: https://github.com/JayKowski/js-testing-prac/graphs/contributors
+[forks-shield]: https://img.shields.io/github/forks/JayKowski/js-testing-prac.svg?style=flat-square
+[forks-url]: https://github.com/JayKowski/js-testing-prac/network/members
+[stars-shield]: https://img.shields.io/github/stars/JayKowski/js-testing-prac.svg?style=flat-square
+[stars-url]: https://github.com/JayKowski/js-testing-prac/stargazers
+[issues-shield]: https://img.shields.io/github/issues/JayKowski/js-testing-prac.svg?style=flat-square
+[issues-url]: https://github.com/JayKowski/js-testing-prac/issues
