+++ files/scripts/caesarCipher.js
@@ -1,17 +1,19 @@
+/* eslint-disable no-nested-ternary */
+
 function caesarCipher(string) {
   const stringArr = string.split(' ');
   const alphabet = 'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ';
   const alphabetArr = Array.from(alphabet);
   const cipheredSentence = [];
-  for (let i = 0; i < stringArr.length; i++) {
+  for (let i = 0; i < stringArr.length; i += 1) {
     let cipheredWord = [];
-    for (let j = 0; j < stringArr[i].length; j++) {
+    for (let j = 0; j < stringArr[i].length; j += 1) {
       if (alphabetArr.includes(stringArr[i].charAt(j))) {
         const currentWord = stringArr[i];
         const charIndex = alphabetArr.indexOf(currentWord.charAt(j));
-        const nextChar = charIndex === 25 ? alphabetArr[0] : 
-              charIndex === 51 ? alphabetArr[26] : 
-              alphabetArr[charIndex + 1];
+        const nextChar = charIndex === 25 ? alphabetArr[0]
+          : charIndex === 51 ? alphabetArr[26]
+            : alphabetArr[charIndex + 1];
         cipheredWord.push(nextChar);
       } else {
         cipheredWord.push(stringArr[i].charAt(j));
@@ -22,7 +24,7 @@ function caesarCipher(string) {
   }
   const completeSentence = cipheredSentence;
   const joinedWordArr = [];
-  for (let k = 0; k < completeSentence.length; k++) {
+  for (let k = 0; k < completeSentence.length; k += 1) {
     joinedWordArr.push(completeSentence[k].join(''));
   }
