+++ files/scripts/capitalize.js
